Fieldname casing option on conversion from IndexDefinition to Azure Index. (#9)
The previous pattern was always to convert to lower case.  This worked fine for auto generated indexes based on Nethereum Event/Function DTO's.   However this causes a problem when using hand coded specific search documents (POCO's).  When writing the doc to the index, unless the casing on the serialization is set specifically, the fields on the doc will not match those on the index and throw an error.