Fix: BlockProgress Repo - GetLastBlockProcessed
Last Block Processed is stored as a string in the DB due to potential size but made returning the "max" value inconsistent with numeric ordering.  The solution is to pad the value written to the repo with zeros.

BlockProgress.LastBlockProcessed Column Length increased from 43 to 100 to be consistent with other BigInteger columns and to provide a consistent padding length.