Provide specific error for SQL Server Data Truncation error
For existing block storage users who may have a BlockProgress table with the old schema and a shorter LastBlockProcessed column length. Catch EF and EFCore exceptions for the UpsertBlockProgressAsync method when data exceeds the max length.  Provide an exception with a message to indicate that the LastBlockProcessed column may not be large enough and states what the column length should be.  Includes the root exception in the inner exception.